add MSP View, add roles, add tests for companyService

+++ FinancialServices/Data/Models/MapingCompanyReport.cs
@@ -16,7 +16,7 @@ namespace Theatre.Data.Models
 
 
 
-        public long ReportId { get; set; }
+        public int ReportId { get; set; }
 
         [ForeignKey(nameof(ReportId))]
         public ReportData ReportData { get; set; }
